fintech-saas/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2024.1
Pillow==10.1.0
